recognize when button was released outside window.
this is a little tricky, not sure the "which" property works the same way
on all the chrome platforms. but, hey, it works for me, right? ugh.
also catch same type of problem with ctrl, which is more rare, but can
happen (e.g.) if a ctrl press activates a button that opens a pop-up dialog.