upgrade to the new storage apis.
rather awkward to handle the asynchronousness. :) the ui on the
options page needed a complete rewrite, naturally. also very slightly
changed the way background sends the options to the infected pages.
there will be a warning when this new mechanism tries to send the
options to the options page, but hopefully chrome will fix that bug.